Grenzel Parity
Removes undivided as a patron from the Grenzel Priest, while also adjusting their skills to be in line with Heartfelt. This doesn't touch their statspread, which is still unique to them.

It also provides ritechalk and a cheele, which they were missing, and an Unidivded(Holy See) amulet, as Imbued is exclusive to the priest/bishop.